add new def of term
The old definition of terms ran into nested recursion errors or
required to be subtypes to exclude a function of arity n being applied
to more than n arguments. Also, any function defined on the old terms
needed to be defined in two mutually recursive parts.

The new definition of terms uses several concepts that conspire
together to resolve these issues and make terms easier to work with.

1. We index every term type with a natural number that keeps track of
its level of nesting.

2. constants and variables are treated as 0-level terms while
functions (of arity n) are treated as n-level terms. Thus, only
functions (and their evaluations as explained below) can have a
positive level.

3. The levels allow us to define partial application (using the `app`
constructor) and side-stepping the issue of nested inductive
types. Only 0-level terms can be applied to (n+1)-level functions to
yield n-level functions.

4. Since application changes (n+1)→0→n instead of n→0→(n-1), this has
the added effect of not allowing application to be applied to
constants or variables, only to functions.

5. Finally, we separated the function (`func` constructor) from
application (using the `app` constructor).